PI controllers for fans and coolant pumps have been added
The new variable WaterTemp has been added to the CarStateType structure. Currently, there is no water temperature sensor, so the controllers cannot be used.
The controllers have been added to the existing temp_ctl() function in power.c. The old code of the function has been commented out in case it is needed in the future, and it can be removed after the controllers are tested. To test the controllers, the FIXEDTEMP value should be changed to 1.